+++ cmd/template/nodepool/provider/capa.go
@@ -7,8 +7,8 @@ import (
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 
-	"github.com/giantswarm/kubectl-gs/cmd/template/nodepool/provider/templates/aws"
-	"github.com/giantswarm/kubectl-gs/internal/key"
+	"github.com/giantswarm/kubectl-gs/v2/cmd/template/nodepool/provider/templates/aws"
+	"github.com/giantswarm/kubectl-gs/v2/internal/key"
 )
 
 func WriteCAPAEKSTemplate(ctx context.Context, client k8sclient.Interface, out io.Writer, config NodePoolCRsConfig) error {